backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
pymysql>=1.1.0
pyjwt>=2.8.0
werkzeug>=3.0.0
apscheduler>=3.10.0
openpyxl>=3.1.0
cryptography>=41.0.0
